requirements.txt
asgiref==3.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.1.6
django-cas-client==1.5.3
django-cas-ng==4.3.0
django-simple-sso==1.2.0
django-uniauth==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itsdangerous==0.24
lxml==4.9.2
numpy==1.20.1
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==2.6.0
python-cas==1.6.0
pytz==2021.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.14
webservices==0.7
django-relativedelta==1.1.2
django-xhtml2pdf==0.0.4
